refactor: move clar into separate directory
