+++ tests/clar/clar.c
