+++ tests/clar/clar_libgit2.c
