+++ tests/clar/clar_libgit2_timer.c
