+++ tests/clar/clar_libgit2_trace.c
