+++ tests/clar/generate.py
